Add Support for generating the application services as typescript not only the type definition
