+++ src/Abp.Web.Api/WebApi/Controllers/Dynamic/Scripting/TypeScriptController.cs
@@ -10,13 +10,24 @@ namespace Abp.WebApi.Controllers.Dynamic.Scripting
     public class TypeScriptController : AbpApiController
     {
         readonly TypeScriptDefinitionGenerator _typeScriptDefinitionGenerator;
-        public TypeScriptController(TypeScriptDefinitionGenerator typeScriptDefinitionGenerator)
+        readonly TypeScriptServiceGenerator _typeScriptServiceGenerator;
+        public TypeScriptController(TypeScriptDefinitionGenerator typeScriptDefinitionGenerator, TypeScriptServiceGenerator typeScriptServiceGenerator)
         {
             _typeScriptDefinitionGenerator = typeScriptDefinitionGenerator;
+            _typeScriptServiceGenerator = typeScriptServiceGenerator;
         }
+
         public HttpResponseMessage Get()
         {
             var script = _typeScriptDefinitionGenerator.GetScript();
+            var response = Request.CreateResponse(System.Net.HttpStatusCode.OK, script, new PlainTextFormatter());
+            response.Content.Headers.ContentType = new MediaTypeHeaderValue("application/x-javascript");
+            return response;
+        }
+
+        public HttpResponseMessage GetCompleteServices()
+        {
+            var script = _typeScriptServiceGenerator.GetScript();
             var response = Request.CreateResponse(System.Net.HttpStatusCode.OK, script, new PlainTextFormatter());
             response.Content.Headers.ContentType = new MediaTypeHeaderValue("application/x-javascript");
             return response;
